Extract lens style from lens editor app
