+++ x-pack/platform/plugins/shared/lens/test/scout/open_in_lens/ui/parallel_tests/agg_based/metric.spec.ts
@@ -187,7 +187,7 @@ spaceTest.describe('Lens open in Lens — agg-based Metric', { tag: tags.deploym
 
     await dimensions[0].click();
     await lens.openPalettePanelFlyout();
-    const colorStops = await lens.getPaletteColorStops();
+    const colorStops = await lens.style.getPaletteColorStops();
     expect(colorStops).toStrictEqual([
       { color: 'rgba(0, 104, 55, 1)', stop: '12000000000' },
       { color: 'rgba(183, 224, 117, 1)', stop: '13000000000' },
