+++ x-pack/platform/plugins/shared/lens/test/scout/open_in_lens/ui/parallel_tests/agg_based/pie.spec.ts
@@ -119,7 +119,7 @@ spaceTest.describe('Lens open in Lens — agg-based Pie', { tag: tags.deployment
     await lens.waitForVisualization('partitionVisChart');
 
     expect(await lens.getChartSwitchType()).toBe('Pie');
-    expect(await lens.getDonutHoleSize()).toBe('Small');
+    expect(await lens.style.getDonutHoleSize()).toBe('Small');
   });
 
   spaceTest('should convert Pie types correctly', async ({ pageObjects }) => {
