+++ x-pack/platform/plugins/shared/lens/test/scout/open_in_lens/ui/parallel_tests/tsvb/metric.spec.ts
@@ -90,7 +90,7 @@ spaceTest.describe('TSVB Metric - Open in Lens', { tag: tags.deploymentAgnostic 
     // Open the metric dimension editor and verify converted palette color stops
     await dimensions.locator('nth=0').click();
     await lens.openPalettePanelFlyout();
-    const colorStops = await lens.getPaletteColorStops(2);
+    const colorStops = await lens.style.getPaletteColorStops(2);
     // Converted color rule from TSVB background_color_rules
     expect(colorStops).toStrictEqual([
       { stop: '10', color: 'rgba(84, 179, 153, 1)' },
